requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-import-export==3.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pypng==0.20220715.0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.4
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
xlrd==2.0.1
xlwt==1.3.0
